# Optimize pwndbg.exception import time