@disconnect3d 2024-01-02
I've also added a lazy import of pkg_resources which accounts for most of the time of this import. Now it went down to:

```
root@pwndbg:~/pwndbg# PYTHONPROFILEIMPORTTIME=1 gdb --batch 2>&1 | grep 'pwndbg\.exception'
import time:       296 |        760 |     pwndbg.exception
root@pwndbg:~/pwndbg# PYTHONPROFILEIMPORTTIME=1 gdb --batch 2>&1 | grep 'pwndbg\.exception'
import time:       301 |        785 |     pwndbg.exception
^[[Aroot@pwndbg:~/pwndbg# PYTHONPROFILEIMPORTTIME=1 gdb --batch 2>&1 | grep 'pwndbg\.exception'
import time:       304 |        866 |     pwndbg.exception
```

However, this did not influence the startup time, because unfortunately `pkg_resources` is imported by `capstone` which it is not so trivial to lazy import at this moment.